refactor: UI 처리를 위해 페이지 분리
